+++ src/command_base.cpp
@@ -56,3 +56,15 @@ std::shared_ptr<Subcommand> CommandBase::getCommand(ArgParser& parser)
 
     return cmd;
 }
+
+bool CommandBase::_strEndsWith(const std::string& fullString, const std::string& ending)
+{
+    if (fullString.length() >= ending.length())
+    {
+        return (0 == fullString.compare(fullString.length() - ending.length(), ending.length(), ending));
+    }
+    else
+    {
+        return false;
+    }
+}