+++ src/write_command.cpp
@@ -18,6 +18,7 @@ bool WriteCommand::parseArguments(ArgParser& parser)
     if (!parseTopics(parser, &m_allTopics, &m_topics))
         return false;
 
+    // -f
     const char* filename = parser.getArg("f");
     if (filename == NULL)
     {
@@ -27,6 +28,7 @@ bool WriteCommand::parseArguments(ArgParser& parser)
 
     m_filename = filename;
 
+    // --lz4 or --bz2
     if (parser.hasArg("bz2") && parser.hasArg("lz4"))
     {
         printf("error: must choose from eithor --bz2 or --lz4\n");
@@ -39,6 +41,37 @@ bool WriteCommand::parseArguments(ArgParser& parser)
     else
     {
         m_compressType = CompressionType::lz4;
+    }
+
+    // --duration
+    if (!parser.hasArg("duration"))
+    {
+        printf("error: please provide --duration 60s or --duration 1m (max 300s / 5m)\n");
+        return false;
+    }
+    std::string duration_limit = parser.getArg("duration");
+
+    if (CommandBase::_strEndsWith(duration_limit, "s"))
+    {
+        int nPos = duration_limit.find("s");
+        std::string dt = duration_limit.substr(0, nPos);
+        m_durationLimit = atoi(dt.c_str());
+    }
+    else if (CommandBase::_strEndsWith(duration_limit, "m"))
+    {
+        int nPos = duration_limit.find("m");
+        std::string dt = duration_limit.substr(0, nPos);
+        m_durationLimit = atoi(dt.c_str()) * 60; // unit: seconds
+    }
+    else
+    {
+        m_durationLimit = atoi(duration_limit.c_str());
+    }
+
+    if (m_durationLimit == 0)
+    {
+        ROS_ERROR("write duration limit must not be 0: %s", duration_limit.c_str());
+        return false;
     }
 
     return true;
@@ -55,6 +88,7 @@ int WriteCommand::run()
 
     TriggerRecordRequest req;
     req.compression_type = (int)(m_compressType);
+    req.duration_limit = m_durationLimit;
 
     boost::filesystem::path p(boost::filesystem::system_complete(m_filename));
     req.filename = p.string();