+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- 2022-02-25 Add write file duration limit. Use `axrosbag write -f xxx.bag --duration 60s/1m`
 - 2022-02-14 Support resampling of message. Use `axrosbag daemon /odom@3hz`.
 - 2022-01-14 Fixed bug of service /axrosbag/write
 - 2022-01-12 Add `axrosbag pause/resume`. It can be used to temporarily stop recording some topics.